+++ src/Types.fs
@@ -192,7 +192,13 @@ module internal TopicPartitionOffset =
             tp.Offset |> Option.map Offset.toKafka |> Option.defaultValue Confluent.Kafka.Offset.Unset
         )
 
-type GetCheckpoint = TopicPartition -> AsyncResult<TopicPartitionOffset, exn>
+type GetCheckpoint = GroupId -> TopicPartition -> AsyncResult<TopicPartitionOffset, exn>
+
+[<RequireQualifiedAccess>]
+module Checkpoint =
+    let key (topicPartition: TopicPartition) = function
+        | GroupId.Random -> Error "Random group id doesn't make sense for checkpoint"
+        | GroupId.Id groupId -> sprintf "%s@%s:%d" groupId topicPartition.Topic.Value topicPartition.Partition |> Ok
 
 //
 // Utilities